Move grammar for 1D CNN to experimental
